+++ src/main/java/barch/mc_extended/Villagers/VillagerPOIs.java
@@ -2,6 +2,7 @@ package barch.mc_extended.Villagers;
 
 import barch.mc_extended.MCExtended;
 import net.fabricmc.fabric.api.object.builder.v1.world.poi.PointOfInterestHelper;
+import net.minecraft.block.Blocks;
 import net.minecraft.util.Identifier;
 import net.minecraft.world.poi.PointOfInterestType;
 
@@ -14,6 +15,8 @@ public class VillagerPOIs {
 
     // gem dealer
     public static PointOfInterestType GEM_DEALER_POI;
+    // miner
+    public static PointOfInterestType MINER_POI;
 
 
     public static void registerPOIs() {
@@ -25,6 +28,9 @@ public class VillagerPOIs {
 
         // gem dealer
         GEM_DEALER_POI = PointOfInterestHelper.register(new Identifier("mc-extended", "gem_dealer"), 1, 10, MCExtended.GEM_TABLE);
+
+        // miner
+        MINER_POI = PointOfInterestHelper.register(new Identifier("mc-extended", "miner"), 1, 50, Blocks.FURNACE);
 
     }
 }